+++ neocaml-menhir.el
@@ -253,6 +253,10 @@ via language injection.
   ;; Final newline
   (setq-local require-final-newline mode-require-final-newline)
 
+  ;; Make URLs and bug references in comments clickable
+  (goto-address-prog-mode)
+  (bug-reference-prog-mode)
+
   (treesit-major-mode-setup))
 
 ;;;###autoload